[msm] split up seperate MSM estimators into individual files
Rationale:
The maximumlikelihood module contained a base class, ml, oom and augmented
estimators adding up to >2k lines of code, which makes it hard to maintain.

Fixes #1260
